# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/portal/login": {
      "post": {
        "operationId": "PortalController_signIn[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sign in; returns every workspace (role × scope) and requires_scope_selection",
        "tags": [
          "portal"
        ]
      }
    },
    "/portal/login": {
      "post": {
        "operationId": "PortalController_signIn[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sign in; returns every workspace (role × scope) and requires_scope_selection",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/v1/portal/select-scope": {
      "post": {
        "operationId": "PortalController_selectScope[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectScopeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activate a workspace; returns active_scope + a token scoped to it (switch without logging out)",
        "tags": [
          "portal"
        ]
      }
    },
    "/portal/select-scope": {
      "post": {
        "operationId": "PortalController_selectScope[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectScopeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activate a workspace; returns active_scope + a token scoped to it (switch without logging out)",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/v1/portal/logout": {
      "post": {
        "operationId": "PortalController_signOut[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/portal/logout": {
      "post": {
        "operationId": "PortalController_signOut[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "portal"
        ]
      }
    },
    "/v1/login": {
      "post": {
        "operationId": "CoreLoginController_signIn[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /login - credentials; returns its_id, name, token, requires_scope_selection, assignments",
        "tags": [
          "core-login"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "CoreLoginController_signIn[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /login - credentials; returns its_id, name, token, requires_scope_selection, assignments",
        "tags": [
          "core-login"
        ]
      }
    },
    "/v1/select-scope": {
      "post": {
        "operationId": "CoreLoginController_selectScope[0]_v1",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectScopeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /select-scope - { role_id, scope_type, scope_id } -> { active_scope, token }",
        "tags": [
          "core-login"
        ]
      }
    },
    "/select-scope": {
      "post": {
        "operationId": "CoreLoginController_selectScope[1]",
        "parameters": [
          {
            "name": "x-csrf-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectScopeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "POST /select-scope - { role_id, scope_type, scope_id } -> { active_scope, token }",
        "tags": [
          "core-login"
        ]
      }
    }
  },
  "info": {
    "title": "Miqaat Identity Federation Service",
    "description": "Central embedded login, federation sessions, RS256 assertions (JWKS), SSO and federation logout. Authentication only - authorization decisions live in the Core Identity Authorization Service.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PortalLoginDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string"
          },
          "identity_type": {
            "type": "string",
            "enum": [
              "ITS",
              "NON_ITS"
            ],
            "default": "ITS"
          },
          "its_id": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "transaction_id",
          "password"
        ]
      },
      "SelectScopeDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "description": "Portal transaction from the login page (CSRF binding)"
          },
          "role_id": {
            "type": "string",
            "example": "6b1f3c2e-8a3d-4a5b-9c1e-2d3f4a5b6c7d"
          },
          "scope_type": {
            "type": "string",
            "enum": [
              "CORE",
              "BUSINESS_UNIT",
              "UTILITY"
            ],
            "example": "UTILITY"
          },
          "scope_id": {
            "type": "object",
            "nullable": true,
            "description": "NULL for CORE, else bu_id or utility_id"
          },
          "audience": {
            "type": "string",
            "enum": [
              "authorization",
              "identity"
            ],
            "default": "authorization",
            "description": "authorization = Core Identity Authorization APIs (e.g. GET /me/permissions); identity = this service (force federation logout)"
          }
        },
        "required": [
          "transaction_id",
          "role_id",
          "scope_type"
        ]
      },
      "PortalLogoutDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "enum": [
              "session",
              "federation"
            ],
            "default": "session"
          }
        },
        "required": [
          "transaction_id"
        ]
      }
    }
  }
}